+++ AutoparkV3/AutoparkV3/Vehicles/VehicleType.cs
@@ -5,11 +5,11 @@ namespace AutoparkV3.Vehicles
     public class VehicleType
     {
         public string TypeName { get; set; }
-        public float TaxCoefficient { get; set; }
+        public double TaxCoefficient { get; set; }
 
         public VehicleType() { }
 
-        public VehicleType(string typeName, float taxCoefficient = 1.0f) 
+        public VehicleType(string typeName, double taxCoefficient = 1) 
         {
             TypeName = typeName;
             TaxCoefficient = taxCoefficient;
